move from home page to header component
